DRY up schema and client tests with shared examples
