+++ gems/smithy/spec/support/contexts/generated_client_context.rb
