+++ gems/smithy/spec/support/contexts/generated_schema_context.rb
