+++ gems/smithy/spec/support/examples/module_examples.rb
@@ -1,0 +1,29 @@
+# frozen_string_literal: true
+
+RSpec.shared_examples 'gem module' do |context|
+  include_context context, fixture: 'weather'
+
+  context 'single module' do
+    it 'has a version' do
+      expect(Weather::VERSION).to eq('0.1.0')
+    end
+
+    it 'requires interfaces' do
+      expect(Weather::Types).to be_a(Module)
+      expect(Weather::Shapes).to be_a(Module)
+    end
+  end
+
+  context 'nested module' do
+    include_context context, fixture: 'weather', module_name: 'SomeOrganization::Weather'
+
+    it 'has a version' do
+      expect(SomeOrganization::Weather::VERSION).to eq('0.1.0')
+    end
+
+    it 'requires interfaces' do
+      expect(SomeOrganization::Weather::Types).to be_a(Module)
+      expect(SomeOrganization::Weather::Shapes).to be_a(Module)
+    end
+  end
+end
