+++ gems/smithy/spec/support/matchers/be_in_documentation_matcher.rb
